Move nested strategy classes to own files
